Finally add dconf autoconfig support!
With this commit, we can finally set up things that use the crappy
GTK dconf binary registry junk. Config files should be named to
match the schema, with config stored as space-separated keys and
values.

The standard schema path for a given app lives in default.dconf-schema
and any relocatable schemas can be defined in $config.dconf-schema.sh
by doing anything necessary to generate the path, and echoing it out.

Lines that are blank or begin with # in config files are ignored.