Add vim-obsession with helper func for zoooom
Vim-obsession adds an :Obsession command to automatically keep track of
open buffers, panes etc and resumes from a Session.vim file with vim -S.

This also adds a small wrapper func, aliased to `vs`, which not only
takes out the hard part of remembering to start :Obsession in a project,
but also keeps the Session.vim file out of the way in ~/.cache/vim.

Usage: `vs some-project` to both start *and* resume a project.