Parse user/repo and branch from Git URL
In the dozen or so -git packages I've got installed, some have multiple
sources including a bunch of local .patch files and some dependencies on
non-Github repositories, but they all have the main source on Github and
listed as the first source in the .SRCINFO file.

I originally planned to loop through, but it's probably unecessary so
instead we can just limit grep to the first match and use that.

We don't need the .git part, or any of the protocol bits which vary
wildly even between the packages I've got installed, so sed strips them
and just returns user/repo, with the branch as a separate value.