Clone AUR repo instead of downloading .SRCINFO
Rather than downloading the package's .SRCINFO file and using ugly Sed
hacks to try and parse the repo/branch out of the first `source` link,
this clones the AUR repo and sources the PKGBUILD directly.

This allows much easier (and more readable) parsing with Bash parameter
expansion, and also frees us up for parsing out the clone alias too,
which will be needed when we want to clone the actual source files.

Also relocates the initial mkdir/cd to create before anything else and
make sure we're always starting in the ~/.cache/aurcheck-git dir.