Update existing repos instead of nuking them
Some packages (like telegram-desktop-git) have a huge amount of history
and take quite a while to download. Nobody likes waiting, so this avoids
having to remove everything each time by checking if the dirs exist.

For the actual source repos we can simply `git fetch` as they're a bare
repository, but we `git pull` the AUR repos to make sure the latest
PKGBUILD gets downloaded to our local copy of the repo.