Remove repo/branch info and section headers
Repo and branch names are arguably useless given all we care about is
the actual package and whether it has updates. In addition, branch will
show as an empty string if it's not custom since removing the Github API
and while we *could* default it to "master", that may not always be the
case. Not worth the API calls just to display in output, and Git doesn't
need it to clone when it's just the default branch, so farewell.

Also removed section headers. A single blank line is more than enough.