Add AUR update check script (#1)
* Add basic update list script

Lists available updates from both the main and user repositories,
seperated under their own headings.

* Add list of all (-git) AUR packages

* Save .SRCINFO for each AUR -git package

* Parse user/repo and branch from Git URL

In the dozen or so -git packages I've got installed, some have multiple
sources including a bunch of local .patch files and some dependencies on
non-Github repositories, but they all have the main source on Github and
listed as the first source in the .SRCINFO file.

I originally planned to loop through, but it's probably unecessary so
instead we can just limit grep to the first match and use that.

We don't need the .git part, or any of the protocol bits which vary
wildly even between the packages I've got installed, so sed strips them
and just returns user/repo, with the branch as a separate value.

* Fix sed. I hate Sed.

The wildcard before `\.git\?` meant that while it worked for some things
to get the branch, others failed completely and really screwed things
up. This is far more verbose and repetitive, but at least it works.

I hope.

* Grab latest commit hash for specified branches

This can be later compared against the installed version, although in a
few cases it'll need some extra work to calculate the "real" version to
prepend to the hash, e.g. 2.1.8r25a89ce4.

* Find default branch from API if unset in SRCINFO

All my local tests seem to result in master, but it's entirely feasible
that some repos might default to develop or some other custom branch.

Better safe than sorry... Though could do with caching. Getting slow.

* Pretty print output from aur-git check

* Organise output into columns

* Skip packages already listed in AUR updates

* Clone AUR repo instead of downloading .SRCINFO

Rather than downloading the package's .SRCINFO file and using ugly Sed
hacks to try and parse the repo/branch out of the first `source` link,
this clones the AUR repo and sources the PKGBUILD directly.

This allows much easier (and more readable) parsing with Bash parameter
expansion, and also frees us up for parsing out the clone alias too,
which will be needed when we want to clone the actual source files.

Also relocates the initial mkdir/cd to create before anything else and
make sure we're always starting in the ~/.cache/aurcheck-git dir.

* Fix branch being overwritten in the loop

* Fix missing newline in AUR Packages section

* Use pkgver() from PKGBUILD to get latest version

Clones the primary source repo into the $package directory and sets
$srcdir so that the pkgver() function in each package's PKGBUILD can
then change into the git repo and do its thing to output version.

* Remove 'latest commit' hash

No longer necessary as the vast majority of version numbers include the
hash anyway, at least now that it's outputting real version numbers.

* Update existing repos instead of nuking them

Some packages (like telegram-desktop-git) have a huge amount of history
and take quite a while to download. Nobody likes waiting, so this avoids
having to remove everything each time by checking if the dirs exist.

For the actual source repos we can simply `git fetch` as they're a bare
repository, but we `git pull` the AUR repos to make sure the latest
PKGBUILD gets downloaded to our local copy of the repo.

* Remove repo/branch info and section headers

Repo and branch names are arguably useless given all we care about is
the actual package and whether it has updates. In addition, branch will
show as an empty string if it's not custom since removing the Github API
and while we *could* default it to "master", that may not always be the
case. Not worth the API calls just to display in output, and Git doesn't
need it to clone when it's just the default branch, so farewell.

Also removed section headers. A single blank line is more than enough.