Add LABEL support to nowplaying blocklet
Rewrites all the player subroutines to return their values rather than
printing directly to STDOUT so that we can print them with the label at
the end, or a simple blank line if there's nothing to output.

Fixes issue where label and/or separator would be displayed when there
are no audio players actually playing anything.