Use Firefox for Nextcloud floating window
Firefox doesn't have a kiosk/app mode so we had to create a new profile
named 'nc' and call it with `--new-instance` while manually specifying a
window class so that we can filter it to toggle floating mode.

The only downside to this is we can only have a single window open at
any given time, but that's not really an issue compared to Chrome nuking
the session and logging out all the fucking time.