Fix transmission not seeing files on launch
Since all the torrents are on my "portable" hdd (don't know why they
call it portable when it requires a 3-pin plug) it takes a moment to
initialise and mount the partition so Transmission often starts up with
all torrents stopped saying the files don't exist.

This is likely because udiskie is started at the same time as everything
else in .xinitrc, so we instead move the responsibility over to i3 so
there should be more than enough of a delay for it to find things.