Finally fix bootup numlock status
Every since the day I installed arch, numlock has always been switched
off during the boot process and it's ridiculously annoying. This finally
fixes it both in a barebones console/tty (with the `numlock` script and
associated .service file) and also i3/X11 with the `numlockx` package.

It would be nice if we could just use the service/script and not need
yet another goddamn piece of software installed, but when applying the
script to `/dev/pts/0` all it does is turn the Numlock light on. Rather
than working as number keys, you then have to press it 3 times to reset
and turn the stupid thing back on again since it goes out of sync.