Shorten output of pacman blocklet
Replaces all the text in the output with a simple "refresh" icon so that
output takes up far less space and fits better on my 2nd monitor.

Also removes the background colour unless it should be inverted, because
the black background looked ugly. I was meant to fix this ages ago but
quite simply never got around to it.