Port highmem and pacman blocklets to Polybar
Adapts the highmem and pacman blocklets I previously wrote for i3blocks.
The pacman script is currently very basic with no support for different
colour levels, but it's enough to show the update count.

For some reason the sudo command to sync pacman doesn't want to work so
for now we're copping out with the basic checkupdates command instead.