Create shorter bar for vertical monitor
Removes the xkeyboard module - because, let's face it, it'll never get
used - and moves the Pacman module to the centre of a new bar created
solely for the right (vertical) monitor. This also uses shorter versions
of the cpu, ram, network, date and volume modules so it can all fit.