Launch pavucontrol when right-clicking volume
This is far from pretty, but it mostly does what we need with the
exceptions being right and left click are swapped compared to the old
behaviour I'm used to with i3blocks, and you can't right-click on the
actual icon unless it's muted. Not perfect, but also not a deal breaker.

We could just define the click handler on the bit appearing first (bar)
and not close the tag, but that somehow prevents mute from working.