Reduce font sizes of the secondary bar
My right monitor is both closer and a lower resolution than the other
monitors running the primary bar, so it can afford to be a bit smaller.

The other upside to this is we can now fit not only the full date, but
even (just about) the volume bar so that's two less duplicates. Huzzah!