Add status of certain services
Adds compton, apache, mariadb, cups and samba status to the bar.

Since this takes up a fair chunk of space on the left monitor, this
creates a new bar for it that removes weather, nowplaying and wpm/trash.
With each monitor now having its own bar, we can remove the MONITOR var.

Both the compton and sysctl scripts are taken from polybar-scripts.