Revert git status output to checkout/reset
As of Git v2.23.0, what used to be a simple checkout/reset command is
now split into a bunch more commands, with git status' output changed to
promote the idiot-friendly switch/restore instead.

It's totally unnecessary, but there you go. For now, checkout still does
what it's always done, so to avoid confusion this simply replaces the
text in the status output and we can ignore the idiot-friendly bullshit.

Switch and restore are supposedly still "experimental" but should the
behaviour of the proper checkout and reset commands ever get fucked up
in favour of this new crap, we can just fork and revert the commits on
the gitster/git@nd/switch-and-restore branch. The original merge commit
can be found at git/git@f496b064fc1135e0dded7f93d85d72eb0b302c22, but
there are so many commits to add it that the message is truncated.

That's not the only commit though, so it may be wise to do a search for
any commits by @pclouds that start with 'switch:' or 'restore:' to make
sure any future commits (such as fixes and later tweaks) are included.