Add custom keyboard config for xorg
Sets the menu button (between R_Super and R_Ctrl) to act as a compose
key to enable typing accented chars and combinations such as menu+o+c
to get a copyright symbol, menu+L+- for £, menu+c+= for € and so on...

Probably don't need to set the layout and model here, but it can't hurt.
Also enables cursor keys when shift and num lock are pressed together.