Fix up the pacman update script
This reverts back to using pacman -Sup instead of checkupdates, with
pacman -Sy to update the repos before checking update count. This way,
we don't need to update repos every time we run `aurupdates` in shell.

Also adds support for yay if it's available, because for some reason the
pacaur command doesn't always work so well in polybar as it did in
i3-blocks. Yay automatically excludes ignored packages, so we don't even
need to grep that stuff out. Bonus! If neither are available, assume 0.