Fix shadow around tray icons
Compton draws a shadow on all sides around the tray container, giving it
an ugly black "glow" separating it from the rest of the bar. This adds
an exclusion for the tray window to solve that issue.

We also exclude Chrome from opacity rules, since it's only used to test.