Replace Chrome with nvidia-smi on startup
Chrome Beta hasn't worked in a month or two. It could possibly be
fixable with a rebuild, but I haven't tried yet. If I recall correctly,
there were issues with requiring outdated dependencies. But it could
just be that it was *built* with older versions that have been replaced.

Instead, this starts up nvidia-smi so we can monitor our GPU fan speed.