Fix vim colour scheme for background transparency
Disables termguicolors because Vim isn't able to "unset" the background
colour as long as it's enabled, and enables the transparent option for
Neodark which needs to be before colourscheme is set.

Also sets some custom colours in Alacritty to fix the ugly grey bars and
cursorline resulting from notermguicolors, and adds my own custom theme
for Vim Airline based on the Neodark, Bubblegym and Wombat themes.