Fix "failed to prepare transaction"
Caused by pacman's inability to satisfy dependencies for something.
Seems this is caused by xorgproto breaking dependencies for a couple
libs this time, but it could be caused by any package later down the
line so it's probably best that we fix it now.

The -Qun method doesn't seem affected and will still list the packages
available for update. We could probably to better than replacing one
line with three by simply using pacman -Qun instead of calling it via
the AUR helpers, which would also remove the need for the final -Sup.

Oh well. This seems to be working for now. Let's see how/if the
aurupdates script is affected (ie if -Sy && -Qun is enough to update).