Remove dependency on gnome-desktop
Both nautilus and gnome-font-viewer depend on gnome-desktop, and neither
of them are actually that good. Nemo is an alternative forked from
Nautilus, but that still depends on dconf and if I wanted my config
managed by a database I'd be using Windows.

Thunar OTOH does all I need, with thumbnails provided by tumbler. The
media tags plugin adds an extra tab to file properties that enables
editing of ID3 tags and probably also EXIF data for photos. Whether I'll
end up using it remains to be seen, but it could be handy eventually.

Font Manager is also a heck of a lot more useful than the GNOME font
viewer which is about as dumbed down as it could possibly get without
being completely useless.