Replace Corebird with Cawbird
The former was abandoned after Twitter ballsed up the API making it
impossible for third party apps to get updates pushed automatically.

Cawbird is a fork that aims to provide similar functionality by polling
for new tweets once in a while which, while not perfect, still beats
using the Twitter website which seems to get worse every year.