Add Ranger config and highlighting
Turns out syntax highlighting in Ranger's file preview pane is as simple
as installing the highlight package. Thanks, @sstutz!

Also customises some of the config options, most notably the column
layout to make it a bit better in wider terminals, and relative line
numbers like we use in Vim to make jumping around a bit quicker.