Use dunst for notifications instead of notify-osd
Like most things Ubuntu/GNOME, notify-osd has absolutely no
configuration whatsoever and only provides the most basic of
functionality when it comes to the notifications displayed.

This adds dunst with some minor tweaks made to the default config so
that notifications with URLs can actually be opened, and isn't limited
to showing ONE notification at a time which—when you first start
Weechat—is the most annoying thing EVER and usually results in pkill.

For the context menu to work it relies on dmenu which I refuse to
install, so for now that option is commented out. Single-URL
notifications can still be opened, but more fiddling might be needed for
those that have multiple actions/URLs. Maybe switch to Rofi? *shudder*