Add zplug install bits
While there's a recommended install script we could use, all it does is
ensure supported versions of git and zsh are installed before cloning
zplug, so we may as well just clone it because git is always installed.

The default would be ~/.zplug, so we use some env vars to use XDG dirs.