Add bullet train theme and (useful) OMZ libs
Most of the nearly-two-dozen "libs" in oh-my-zsh are there purely to
provide functions for theme support and don't really provide any useful
features so rather than including all of it, we just take what's really
necessary to get the shell just the way we want it.

Doing it this way means we also don't need to source the dreaded
key-bindings.zsh lib which has always hijacked the bindkeys mode
resulting in a broken vi-mode plugin. Without that script, it seems
Alacritty is more than capable of doing it itself—no vi-mode needed!