Add and update custom themes/plugins from OMZ
Rather than keeping a copy of the git-flow-completion script we can just
load it as a plugin, and same goes for `k` vs the git submodule.

For the theme we clone the repo so that the Node and Elixir versions are
only shown if we're in a project or directory that contains the relevant
files/folders. E.g. we don't need to see elixir version in PHP projects.