Add environment vars from OMZ
This also has the wonderful side effect of cleaning up a bunch of
duplication that was affecting $PATH in oh-my-zsh. At the same time, we
also move Go's files into ~/dev/go and binaries in ~/.local/bin/go.

Not entirely sure we need $GOBIN in $PATH, got a funny feeling
$ZPLUG_BIN covers it and is included automatically by zplug.