Fix k plugin and dir listing aliases
Turns out because plugins are all loaded at once as the last step in the
process, our `k="k -h"` alias blocks supercraptree/k from loading the k
function itself, resulting in command not found when trying to run it.

Instead of aliasing k, this updates to add the -h arg to our ls aliases.