Zplug (#2)
* Add zplug install bits

While there's a recommended install script we could use, all it does is
ensure supported versions of git and zsh are installed before cloning
zplug, so we may as well just clone it because git is always installed.

The default would be ~/.zplug, so we use some env vars to use XDG dirs.

* Add bullet train theme and (useful) OMZ libs

Most of the nearly-two-dozen "libs" in oh-my-zsh are there purely to
provide functions for theme support and don't really provide any useful
features so rather than including all of it, we just take what's really
necessary to get the shell just the way we want it.

Doing it this way means we also don't need to source the dreaded
key-bindings.zsh lib which has always hijacked the bindkeys mode
resulting in a broken vi-mode plugin. Without that script, it seems
Alacritty is more than capable of doing it itself—no vi-mode needed!

* Add and update custom themes/plugins from OMZ

Rather than keeping a copy of the git-flow-completion script we can just
load it as a plugin, and same goes for `k` vs the git submodule.

For the theme we clone the repo so that the Node and Elixir versions are
only shown if we're in a project or directory that contains the relevant
files/folders. E.g. we don't need to see elixir version in PHP projects.

* Add environment vars from OMZ

This also has the wonderful side effect of cleaning up a bunch of
duplication that was affecting $PATH in oh-my-zsh. At the same time, we
also move Go's files into ~/dev/go and binaries in ~/.local/bin/go.

Not entirely sure we need $GOBIN in $PATH, got a funny feeling
$ZPLUG_BIN covers it and is included automatically by zplug.

* Add custom git/gitflow aliases and functions

This incorporates a basic structure for categorising aliases and various
helper functions into ~/.config rather than the old ~/.zsh-alias dir.

We're starting by migrating OMZ's git plugin to zplug, along with most
aliases for that and git flow. Since the latter now has its own file,
we'll opt to load the git-flow-completion plugin from that instead.

* Port over remaining lines from .zshrc

This moves everything that was left in the old zshrc including aliases
and any oh-my-zsh plugins that are still used. Some effort has been made
to organise things slightly better and prune some bits that never get
used, but there are probably still some improvements that could be made.

* Add history substring search to up and down keys

* Reorganise things to clean up main .zshrc file

* Add some extra bits

Adds fzf completion and enables syntax highlighting for ZSH, which needs
to be loaded before the history substring search for whatever reason.

Also adds a bit that tells ZSH where to put its zcompdump files, which
will be a nice thing to keep out of the home directory at long last...

* Fix zplug script failing to source files

* Replace history-substring-search with the old way

I don't like the fuzzy search.

* fix config loading

* Replace history substring and fix keybindings

Adds a whole ton of key bindings to get the functionality we've come to
expect from keys such as Home/End, Up/Down, Delete and so on.

* Replace OMZ with Zplug

* Manually implement (most of) lib/directories.zsh

There are a couple bits we don't use, and most of the ls aliases
conflict with ones we're trying to add, so this gets more specific.

* Fix k plugin and dir listing aliases

Turns out because plugins are all loaded at once as the last step in the
process, our `k="k -h"` alias blocks supercraptree/k from loading the k
function itself, resulting in command not found when trying to run it.

Instead of aliasing k, this updates to add the -h arg to our ls aliases.