Fix software that depends on storage drive
If udiskie takes longer than normal to mount the drive (or it isn't
plugged in) then nextcloud, transmission and feh will all fail in some
way or another. By delaying the startup of those things until the data
mount becomes available, most of those issues are solved.

Transmission still seems to have issues with half the torrents needing
to be started manually every time it starts, but that's probably because
transmission just sucks. That said, it's still better than it was before.