Fix regressions in previous commit
The cleanup broke a few things, notably changing to the dotfiles root
instead of stowable dir, causing 'bootstrap' etc to be wrongly stowed.

When the `depends_on` call was replaced, it also meant the script would
simply be killed if stow didn't exist. This fixes it by "or-ing" into
the install switch rather than exiting the script entirely.

Finally, this also fixes an issue where applying the `action` and `flag`
values too early meant "Stowing" or "Restowing" would not show in output
past the first package, and may or may not've affected the stow command.