Move dconf/gsettings to standalone script
No matter what I try, and even with X started in another tty and dbus
service started beforehand, gsettings simply refuses to apply any
configs from a secondary tty on first boot. I can't be bothered trying
anything else to workaround gnome's nonsense, so this moves it to a new
script called `apply-gsettings` which will run when a terminal's opened.

To avoid undesired repeated runs, we use .installed.gsettings to track
whether the script has been fired or not. That has no effect other than
when called from the .zshrc as that is the only place that checks it.