Fix ps_mem-related journal spam
One invocation of ps_mem with sudo causes 6 lines in the journal.
Every 5 seconds. With no sleep, that's over 103,000 lines. Daily.

This commit updates highmem to run in a loop instead, using polybar's
tail mode so that we can simply call the script with sudo, not ps_mem.
Just to simplify things, it's moved into .local/bin as 'memwatch'.

---

Also updates the update checker frequency because every 10 minutes is
simply too much. I probably set it that way so I could be lazy with repo
pulls when installing random packages, but it ain't worth it.

A source on IRC suggests checking AUR every 5 minutes however could
result in bans. This wouldn't be ideal and could happen accidentally
should I be doing it manually while installing/checking other things.