More UI plugins (and replacements)
This adds telescope to provide the ctrl-p/g bindings that previously
triggered fzf (and the dedicated ctrlp plugin prior to that). Config
isn't yet perfect, may need to shorten paths and use fzf-native to
potentially improve the sorting of results, but we'll see later.

Also adds Vista to replace Tagbar with the added bonus of LSP support.
May need configuration to use LSP for PHP etc, haven't tested that yet.

Peekaboo is copied directly from vim config (for previewing "/@)