Setup Telescope to use rg with fzf-native sorting
Live grep is not meant to work like `:Rg`, so this adds a custom "scope"
wrapping the `find_files` builtin and configures everything to use the
telescope-fzf-native extension so we get filtering more like fzf.

This also fixes a long-running issue with `:Rg` where—in an effort to
show *hidden* files—it would also include results from the .git dirs.